Abstract migrator part of `getMaxiBlockAttributes`
